Add contributing docs when building component from scratch (#811)
# Pull Request

## 🤨 Rationale

As we've work on the new wafer map component I want to collect any
information we share with that team in our documentation to make the it
easier for future contributors to get up to speed.

## 👩‍💻 Implementation

Add a section in the `nimble-components` CONTRIBUTING.md that covers
building component from scratch, not leveraging FAST Foundation or
composing other Nimble components. Link to FAST's docs on templates.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [X] I have updated the project documentation to reflect my changes or
determined no changes are needed.