Fix intermittent drawer test (#819)
# Pull Request

## 🤨 Rationale

Resolves #816 

## 👩‍💻 Implementation

I wasn't able to determine why the previous implementation of
`completeAnimationAsync` in the tests was causing intermittency, but I
think the updated implementation is better regardless. Rather than
polling for the `animating` CSS class to be removed from an element to
know that the animation is complete, I am now adding an `animationend`
event listener to know when the animation is complete.

## 🧪 Testing

- Ran the tests a number of times locally (although this always passed
with the old implementation as well)
- Ran the pipeline a number of times

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.