Adjust control height based on label presence (#808)
## 🤨 Rationale

Fixes #480

## 👩‍💻 Implementation

If we no longer explicitly set the control height in the CSS of the
text/number field, then the height is calculated the way that we want
(i.e. 48px with a label and 32px without). I confirmed that showing the
error string still doesn't affect the calculated height, which is
correct.

## 🧪 Testing

Tested in Storybook

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.