Create WaferMap skeleton in nimble repo (#818)
# Pull Request

## 🤨 Rationale

Create WaferMap skeleton in nimble repo. We want to rewrite all the
WaferMap component.

https://dev.azure.com/ni/DevCentral/_workitems/edit/2208924


## 👩‍💻 Implementation

I added the simple, empty files for the WaferMap component. For the
moment, it is just the structure of the files and the interface of the
component. The interface will be updated in the future, for the moment
we want to keep it as before (we just delete the parts that we don't
need, but keep the old structure).

## 🧪 Testing

Added one test for the component and it pass. 

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Jonathan Meyer <jonathan.meyer@ni.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>