GitHub workflow permissions test (#901)
# Pull Request

## 🤨 Rationale

This is a test to see if switching to workflow permissions (the
permissions key in the workflow) from a personal access token (the
previous secrets.GITHUBPAGESDEPLOYTOKEN) will work correctly with
beachball publish.

## 👩‍💻 Implementation

Enabled the `contents: write` permission on the workflow.

## 🧪 Testing

Doing it live by publishing a patch change to blazor (fewer external
users than components or angular). Will be prepared to revert if needed.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.