Created Prerendering Module which prepares information for each die to be ready for rendering  (#835)
# Pull Request

## 🤨 Rationale

The wafer map component is a skeleton which will be filled with the
required functionality step by step. This is the step where the
prerendering module and the data manager module functionality is added.

## 👩‍💻 Implementation

The Prerendering module uses the scales built and the dimensions
calculated in the computations module to create rendering ready dies
with the specific position, color, label, and opacity attributes.

The data manager module will tie up the prerendering and calculations,
instantiating them and providing the link between the main component and
the two modules while revealing only the needed interfaces to the rest
of the component.

## 🧪 Testing

Created automated tests for the two modules created.
Tests are passing locally and in the branch build ci

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.