Alphabetize component status table and update issue link (#903)
# Pull Request

## 🤨 Rationale

The nimble-table entry in the status table was out of order which led me
to think we had forgotten to update its status. It also pointed to an
old table issue

## 👩‍💻 Implementation

Update the spreadsheet order, link to the table project instead of a
table issue, and regenerate the status table markdown. Manually remove
diffs for anchor which were already present in the spreadsheet since
they'll be committed with #832

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [X] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>