Basic TanStack integration with nimble-table (#902)
# Pull Request

## 🤨 Rationale

This PR is part of the effort for #840.

This PR pulls in the TanStack dependency to nimble-components and
updates the `nimble-table` to render basic data by leveraging TanStack.

Things that are temporary in this PR:
- Determining the set of columns by looking at the keys in the data
- This will ultimately be determined by column definitions provided to
the table (likely slotted as children of the table)
- Turning the data into `string[][]` type for rendering
- This will ultimately use view templates provided by the column
definitions
- The entire template, including styling
- This will ultimately be updated to have additional `nimble-*` elements
that make up the table. Some examples may be `nimble-column-header`,
`nimble-row`, and `nimble-cell`.
- The template will ultimately need to be setting appropriately ARIA
attributes

## 👩‍💻 Implementation

- Add a dependency on the latest version of TanStack
- Use the TanStack APIs to create a table using the basic
`getCoreRowModel()` function. Currently the only relevant information
we're giving TanStack is the data, but as development continues on the
table, the same patterns can be used for other information, such as
sorting & grouping state.
- Create an interface named `TableData` that describes the required type
of `TData` passed to the table
- Provide a property on the `nimble-table` that allows setting data,
which is rendered in the table

## 🧪 Testing

- Manually tested via storybook
- Wrote unit tests, primarily focused around being able to update the
data in the table
- As much/all of the template is likely to change, I created a page
object for the table in hopes that it will help us avoid having to
continually update tests as we iteratively update the template
- I had to update `karma.conf.js` to get TanStack to work in the unit
tests

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>