Changed wafer map rendering die opacity to rgba color alpha channel (#909)
# Pull Request

## 🤨 Rationale

The opacity atribute of the rendering die was combined with the color in
the alpha cannel by the rendering function in the previous wafer map
implementation, but we decided to move it in the prerendering module

## 👩‍💻 Implementation

Created a fast-colors ColorRGBA64 object from the scale provided value
and set it's alpha channel to the specific opacity.

## 🧪 Testing

Existing tests were updated to check for the changed property.

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.