Created wafer-map rendering module (#913)
# Pull Request

## 🤨 Rationale

This pull request contains the code which renders the nimble-wafer-map
component.

## 👩‍💻 Implementation

The wafer map outline is drawn using and SVG path. This piece of code is
hardcoded in the template.ts file. The circle scales to the correct size
according to its parent container.
The dies are getting rendered by calling the RenderingModule.drawWafer()
method.
For the initial implementation the size of all the containers are
hardcoded to 500x500 pixels. The dynamic sizing of the wafer-map
component will be covered in a different user story,

## 🧪 Testing
A new `wafer-map-matrix.stories.ts` file was added which contains
scenarios for the chromatic tests.
Please note that as the size of the wafer maps are currently fixed,
instead of one big matrix scenario covering all the possible
permutations of the available inputs, 4 different scenarios were created
as follows:
- Wafer Map Theme Matrix - covers scenarios that shows how the wafer
maps outer circle color adapts to the changing theme color.
- Wafer Map Dies And Orientation Test - covers simple die rendering and
notch orientation changes
- Wafer Map Die Label And Color Scale Test - covers the show/hide die
label and color scheme changes
- Wafer Map Die Orientation Test - covers the die quadrant orientation
changes
Once the dynamic size functionality gets implemented, these tests will
be refactored into a single test covering all the possible scenarios.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: Natan Muntean <natan.muntean@ni.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>