Separate github-pages deploy job (#941)
# Pull Request

## 🤨 Rationale

Changes the GitHub deploy step to use the new [GitHub Pages deploy
actions](https://docs.github.com/en/pages/getting-started-with-github-pages/configuring-a-publishing-source-for-your-github-pages-site#creating-a-custom-github-actions-workflow-to-publish-your-site).
Fixes #679

## 👩‍💻 Implementation

Removed the existing ruby-based approach and switched to the workflow
actions. Used a dedicated job as recommend in the [deploy pages
usage](https://github.com/actions/deploy-pages#usage).

## 🧪 Testing

Doing it live

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.