Fix tab keyboard focus style (#948)
## 🤨 Rationale

Fixes #949.

The keyboard focused styling of the tab did not exactly match the design
spec. Specifically, the extra thin line above the thicker selection line
extended all the way to the edge of the tab instead of being 3px short
on each side.

## 👩‍💻 Implementation

Changed the CSS width to subtract the 6 pixels.

## 🧪 Testing

Storybook

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.