Update to latest versions of @ni/eslint-config-* (#955)
# Pull Request

## 🤨 Rationale

It's always good to be on latest version of our dependencies. This pulls
in some updates to NI's javascript styleguide latest versions which add
a few new rule configurations.

The relevant PRs were:
- https://github.com/ni/javascript-styleguide/pull/105
- https://github.com/ni/javascript-styleguide/pull/117

## 👩‍💻 Implementation

1. Update versions of `@ni/eslint-config-*` to latest
2. Add missing `npm run format` command for `@ni/nimble-tokens`
3. Run `npm run format`
4. Run `npm run lint` and fix remaining violations manually

Most of the fixes are to spacing around braces.

## 🧪 Testing

Relying on pipeline.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.