Add ability to specify a field within the table's data to use as a row's ID (#953)
# Pull Request

## 🤨 Rationale

Implements the row ID logic described in the table's readme, as modified
in #937

## 👩‍💻 Implementation

- Added `id-field-name` attribute
- Added ID validation when the `data` property updates or when the
`id-field-name` updates
- Added a distinction between the `data` on the table and the `rows` on
the table so that the table could render no rows when necessary
- Added `validity` property and `checkValidity()` function

## 🧪 Testing

- Manually tested in Storybook, including dynamically changing the
configuration of the table in dev tools
- Wrote unit tests
- Note: There are no unit tests that validate that a row's ID is the
expected value because the ID is not yet part of the public API of the
table. I expect us to add additional ID tests once we have a public API
that exposes row IDs. However, this is part of what I manually tested in
Storybook.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.