Add Angular support for table row IDs and validity (#961)
# Pull Request

## 🤨 Rationale

Add Angular support for changes made in #953, which includes:
- `id-field-name` attribute on the table
- `validity` property on the table
- `checkValidity()` function on the table

## 👩‍💻 Implementation

- Add items listed above to the Angular directive for the table
- Export the `TableValidity` type from the directive

## 🧪 Testing

- Add unit tests for the directive changes
- Note: I decided to add a `describe` section in the directive unit
tests for "validity", which tests `validity` and `checkValidity()`. My
expectation is that we won't attempt to write Angular tests for every
case that could make the table invalid, but I wanted to put a few tests
in that would sanity test that we are correctly calling into the native
element to get the values for that property and function.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.