Add zoom functionality to wafermap (#923)
# Pull Request

## 🤨 Rationale

The wafer map needs to have a zoom functionality. This is provided by
the d3 library.

## 👩‍💻 Implementation

Added a d3 zoom event on the canvas that modifies the transform of the
zoom-container.

## 🧪 Testing

Tested the change in the storybook and added test for where we simulate
a zooming events.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: ArinLucaNI <119854973+ArinLucaNI@users.noreply.github.com>
Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>
Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>