Implement TableColumn and TableColumnText (#947)
# Pull Request

## 🤨 Rationale
- #876
- #860 
- #850 
- #875 

## 👩‍💻 Implementation

This PR introduces the APIs that will allow us to create custom column
configurations for the `Table`. Additionally, I'm introducing the first
concrete column implementation, `TableColumnText`, as this is the most
common column display.

*_Note_: This PR does _not_ include consuming the `validateCellData` API
of the `TableColumn`. There is some more discussion to be had before we
know exactly how to handle this, and there doesn't seem a strong reason
to gate this PR on figuring that out. I will be filing an issue shortly
after the completion of this PR to track that work.

## 🧪 Testing

Added unit tests, and updated `Table` story matrix to use
`TableColumnText`, exercising its API.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: mollykreis <20542556+mollykreis@users.noreply.github.com>
Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>