Update table styles (#969)
# Pull Request

## 🤨 Rationale

Update the styling of the table:
- Fix background color of rows on Color theme
- Use uppercase text transform on column headers

## 👩‍💻 Implementation

I moved the styling of the row background color from within the row
styles into the table styles so that I could use a `::before` styling on
the row without setting the `position` of the row's `:host` element (see
discussion in
[here](https://github.com/ni/nimble/pull/963#discussion_r1070174660)).

I added an `uppercase` text transform to the column headers since
@nate-ni and @RickA-NI's research indicated that this would not cause an
accessibility or localization issue.

## 🧪 Testing

- Looked at the table in all three themes in storybook.
- Performed basic testing with the Window's narrator that column headers
were announced correctly

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.