Fix table row ids (#980)
# Pull Request

## 🤨 Rationale

After merging #953, I realized that the code to return an empty string
from `getRowId` wasn't causing TanStack to generate an ID. Instead, it
was using an empty string as the row ID. Therefore, when `id-field-name`
wasn't set on the table, every row had the same record ID. I'm not sure
how I missed this because I know I manually tested it, but something
else must have been different in my branch at that time. Regardless, I
fixed the issue and also added a bit more code in order to be able to
write auto tests for record IDs.

Additionally, from an offline conversation, we decided to standardize on
`recordId` rather than `rowId`, so I renamed the validation fields in
this PR as well since I was touching that code.

## 👩‍💻 Implementation

- Add/remove the `getRowId` option on TanStack when `id-field-name`
changes.
- Update the table validator to allow empty string as the value of a
record ID
- Update the field names on `TableValidity` to refer to the `recordId`
rather than the `rowId`
    - Update docs & tests with this change
- Add `record-id` attribute to the table row
- This allowed me to write auto tests. It should also be used in future
features, such as adding an action menu.

## 🧪 Testing

Added auto tests for the table that verify that the correct `record-id`
is set on the `nimble-table-row` elements

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.