Improve rendering speed and fix text alignment issue (#967)
# Pull Request

## 🤨 Rationale

The initial rendering module proved to cause significant slowdowns (10<=
FPS) when it came to rendering high number of dies on the wafer area.
This slowdown got even worse when the values were rendered on each
individual die. This PR provides some improvement for that by:
- rendering text when the dies are of (or zoomed to) a certain size
(this will prevent text from rendering when it's not of a legible size.
- use of `context.fillStyle` changes were reduced (MS recommends
changing this as rarely as possible when drawing on the canvas as it
causes slowdown).
- Split the drawWafer() functionality to two separate functions:
`renderDies()` and `renderText()`

This PR also contains a fix to align the rendered text to the middle of
each die and always compress it so it always leaves 20% (10% each side)
padding in the die.

Reduced the size of the notch.

## 👩‍💻 Implementation

Optimizations: 
- Sorted the the DieRenderInfo[] array by color (fillStyle) so we only
swap the fillStyle of the canvas context when it's done rendering all
the dies of a certain color.
- Moved the text rendering into a separate function which is called
after the plain dies (rectangles) are rendered as in the previous
implementation, the fillStyle of the canvas context had to be swapped
from color to white (and vice versa) every time we rendered a die with
text. Without this split, the first optimization point is useless,
- Instead of rendering the text every time when the wafer map gets
rendered, a new condition was introduced which allows the text to render
only if the die has a certain size.

Updated the drawing path of the SVG so now the circle notch is
proportional to the full circle. (also fixes the issue where the
rendered dies cut into the notch)

Fixed issue which caused the text to exceed the limits of a die. Now the
text can take up a maximum of 80% of the dies width and it's aligned to
the center of the die leaving 10% padding on each side,

## 🧪 Testing

Performance:
Tested it manually using Chrome's dev tool to verify browser FPS.
Chromatic tests will be updated to match the current circle notch.

Performance test results

**Only Dies(no string with values):**

| Data Points  | Current Render Time (ms) | Improved Render Time (ms) |
| ------------- | ------------- |------------- |
| 1k  | 51.70  | 49.70 |
| 10k  | 92.20  | 82.80 |
| 100k  | 453.80 | 402.19 |
| 200k  | 835.40  | 764.59 |
| 500k  | 2278.19  | 1851.30 |

**Dies with value string:**

| Data Points  | Current Render Time (ms) | Improved Render Time (ms) |
| ------------- | ------------- |------------- |
| 1k  | 55.70  | 51.70 |
| 10k  | 153.19  | 126.09 |
| 100k  | 1412.90 | 812.80 |
| 200k  | 4496.80  | 1629.19 |
| 500k  | 11538.19  | 4227.69 |

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

Co-authored-by: Milan Raj <rajsite@users.noreply.github.com>