Menu button updates for table (#986)
# Pull Request

## 🤨 Rationale

Some changes need to be made to the menu button before it can be used
within the table for the action menu. Specifically, those changes are:
- adding a `beforetoggle` event
- having the menu button find the slotted menu when it is nested within
multiple slots

See #956 for more details.

## 👩‍💻 Implementation

- Add `beforetoggle` event to the menu button that is fired before the
menu opens or closes
- Add Angular & Blazor support for the `beforetoggle` event
- Change the logic for finding the slotted menu. We can no longer use a
filter in the template to find slotted elements with a role of `menu`
because this won't match `slot` elements. Instead, the template assigns
all slotted elements to the `slottedMenus` array, and the getter for
`menu` finds the menu, which may be nested within `slot` elements.

## 🧪 Testing

- Manually tested that the `beforetoggle` event handler works correctly
in Blazor & Angular
- Added auto tests for the `beforetoggle` event works correctly for the
web components
- Basic manual testing in Storybook that the component still behaves as
expected
- Added tests that the focus & menu opening behavior works correctly
when the menu is nested in an additional slot

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.