Fix table row rendering in Safari (#998)
# Pull Request

## 🤨 Rationale

Fixes #984 

## 👩‍💻 Implementation

As @rajsite and @msmithNI suggested in the linked issue, use
`this.$fastController.isConnected` instead of `this.isConnected` in the
table cell's `cellTemplateChanged()` method.

## 🧪 Testing

I manually tested in Safari in Storybook and the Angular example app.
Malcolm tested in Playwright's Safari.

Auto tests would be covered by #990.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.