Create setData() function on nimble-table and remove data property (#993)
# Pull Request

## 🤨 Rationale

Fixes #958 

The current `data` property on the `nimble-table` relies on FAST's
`@observable` behavior for arrays to detect updates. This can cause some
unexpected behavior for clients, including having to make a shallow copy
of their data array in some cases. To avoid these oddities and to keep
the `Table` class from having to hold on to its own copy of the data,
the `data` property is being removed and replaced with a `setData`
function that always performs a shallow copy of the data and reassigns
it to TanStack.

## 👩‍💻 Implementation

Web Component
- Remove `data` property
- Add `setData()` function that contains much of the logic the removed
`dataChanged()` function performed
- Perform a shallow copy of each row's data when creating
`TableRowState` instances to ensure that the `@observable` handler for
the `TableRow` was executed when `dataRecord` changes.
- Introduced a `canRenderRows` property that tracks whether the rows
should be rendered in the table based on the table's validity. This
allowed the table code to always update TanStack with the data
(previously it gave TanStack an empty array when the data was invalid)
and avoid having to keep a separate copy of the data within the `Table`
class.

Angular
- Remove `data` property
- Add `setData()` function
- Add `data$` property that encapsulates the logic to call `setData()`
when an Observable emits a new value

Blazor
- Update the interop code to call `setData()` on the web component
rather than assigning to the `data` property
- Did not expose a `setData()` method because it would only wrap a line
of code assigning to the Blazor Table's `Data` property

## 🧪 Testing

- Created auto tests for component and Angular changes
- Manually tested Angular and Blazor example apps
- Updated matrix tests to have a play function that adds data to the
table

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>