Table virtualization (#966)
# Pull Request

## 🤨 Rationale

Related Issue:
- #865 

## 👩‍💻 Implementation

Uses TanStack Virtualizer as mentioned in the spec, and as done in [the
prototype
branch](https://github.com/ni/nimble/tree/table-custom-columns/packages/nimble-components/src/table).
Major differences from the prototype:
- Prototype and TanStack Virtual examples set the virtualizer's "total
size" (height of all rows) on the row container, and added translate
transforms on each individual row. We noticed some flashing during
scrolling with that approach in Firefox. Instead, this version sets the
height of all rows on a sibling div, and applies a single translate
transform on the row container. This is more similar to what
Perspective/regular-table do, and it got rid of the flashing in Firefox.
- Prototype used a `ResizeObserver` in the Nimble table code. TanStack
Virtualizer is already using a ResizeObserver internally (due to us
using its provided `observeElementOffset` and `observeElementRect`
functions as inputs in our `VirtualizerOptions` object), so I concluded
we didn't need a separate resize observer.

## 🧪 Testing

- Added a couple of autotests dealing with virtualization (that the # of
rows rendered is less than the total; that we can scroll the element to
view the end rows; that the # of rendered rows increases when the
element height increases)
- Updated [table Storybook to add a 2nd option for table
data](https://60e89457a987cf003efc0a5b-wjfhtkxgzq.chromatic.com/?path=/story/table--table&args=data:LargeDataSet),
a data set of 10,000 rows

## ✅ Checklist

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.