Table Storybook setData workaround for Safari (#1009)
# Pull Request

## 🤨 Rationale

Fixes #1007 In debugging I found that when the storybook template is
rendered the custom element is already registered but the nimble-table
instance does not seem to be upgraded yet. Manually calling
[`customElements.upgrade`](https://developer.mozilla.org/en-US/docs/Web/API/CustomElementRegistry/upgrade)
didn't seem to help.

## 👩‍💻 Implementation

Tried waiting on `customElements.whenDefined('nimble-table')` and that
seemed to work in Safari. Still not 100% clear what the root problem is
though... `¯\_(ツ)_/¯`

## 🧪 Testing

Reproduces the original issue and tested the fix on Safari 16.3

## ✅ Checklist

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.