Update instructions for importing icons (#1022)
# Pull Request

## 🤨 Rationale

A developer is [planning to import a new
icon](https://dev.azure.com/ni/DevCentral/_git/Skyline/pullrequest/431242?discussionId=4111723&_a=files&path=%2FWeb%2FWorkspaces%2FTestInsights%2Fsrc%2Fassets%2Ficons%2Fsave-icon.svg)
and asked for instructions. Brandon noticed that the existing
instructions didn't mention that the source of truth for the .ai file is
in Perforce.

## 👩‍💻 Implementation

Add steps describing location of file in Perforce and policy of copying
file to nimble-tokens to the relevant CONTRIBUTING.md.

## 🧪 Testing

N/A

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [X] I have updated the project documentation to reflect my changes or
determined no changes are needed.