Remove explicitly calling out 'null' in type of table's idFieldName property (#1026)
# Pull Request

## 🤨 Rationale

After some discussion within the team, we decided that it would be best
not to explicitly call out `null` as a possible value for `idFieldName`
even though there are code paths that could cause it to be set to `null`
(e.g. calling `table.removeAttribute('id-field-name')`). The rationale
for this decision is that all our string attribute could technically be
null and we don't want to call `null` out on all the types or have
inconsistency in our typing.

Instead, I created the following FAST issue:
https://github.com/microsoft/fast/issues/6630

## 👩‍💻 Implementation

- Remove `| null` from the type of `idFieldName` on the table
- Modify the `TableValidator` class to handle `null` without explicitly
checking for it. This is done using a `typeof idField !== 'string'`
check now.

## 🧪 Testing

- Ran auto tests
- Manually tested in dev tools that calling
`removeAttribute('id-field-name')` on the table does not cause a
validation error

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.