Make the click area of the breadcrumb-item match the link area (#1031)
# Pull Request

## 🤨 Rationale

Fixes #557 

## 👩‍💻 Implementation

The issue was that `onClick` in `RouterLinkWithHref` was called when any
part of the breadcrumb item was clicked, even if it was the separator or
some padding that was clicked. To fix the issue, I:
- provided an override for `onClick` so that the `click` `@HostListener`
of `RouterLinkWithHref` would do nothing
- provided my own `click` `@HostListener` in the nimble directive that
calls `onClick` on the base class only if the anchor within the
breadcrumb item was clicked
- in order to have reference to the element within the directive, I had
to implement a constructor. Because of the way `LocationStrategy` works,
I needed to inject the `Injector` and dynamically get the values for the
`Route`, `ActivatedRoute`, and `LocationStrategy` rather than have them
injected directly as the arguments of the constructor.

## 🧪 Testing

- Manually tested the SLE example app with the built angular package
- Wrote new auto test -- verified it failed before my change and passes
after my change

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [ ] I have updated the project documentation to reflect my changes or
determined no changes are needed.