Wafer Map  - codebase refactor & zoom handling (#1005)
# Pull Request

## 🤨 Rationale

This PR implements an event handling architecture complete with a Zoom
handler. The PR also performs an extensive refactor of the already
existing code base -> index file and types

The PR contains "TODO" items and references to a [HoverHandler
](https://dev.azure.com/ni/DevCentral/_boards/board/t/OptimalPlus%20Dokker/Work%20Items/?workitem=2211981)module
- this module exists, however it was removed from the current Wafermap
version as it is still a work-in-progress and does not meet yet the
required coding/usability standards. Rather than submitting partial
code, we decided to do so in a future PR. This is expected to happen
towards the end of this week.

## 👩‍💻 Implementation

<!---
Describe how the change addresses the problem. Consider factors such as
complexity, alternative solutions, performance impact, etc.

Consider listing files with important changes or comment on them
directly in the pull request.
-->

- In this PR, a lot of changes were applied to the existing Wafer Map
code-base to improve code readability, modularity and ease of
maintenance (index.ts, types.ts).
- A top level EventHandler was added
- A ZoomHandler was added

## 🧪 Testing

<!---
Detail the testing done to ensure this submission meets requirements. 

Include automated/manual test additions or modifications, testing done
on a local build, private CI run results, and additional testing not
covered by automatic pull request validation.
-->
- Manual testing has been done. There are no changes to existing
functionality.

## ✅ Checklist

<!--- Review the list and put an x in the boxes that apply or ~~strike
through~~ around items that don't (along with an explanation). -->

- [x] I have updated the project documentation to reflect my changes or
determined no changes are needed.

---------

Co-authored-by: ArinLucaNI <119854973+ArinLucaNI@users.noreply.github.com>
Co-authored-by: arinluca333 <arin.luca1@gmail.com>
Co-authored-by: Jonathan Meyer <26874831+atmgrifter00@users.noreply.github.com>
Co-authored-by: Jesse Attas <jattasNI@users.noreply.github.com>
Co-authored-by: elizaolariu <eliza.olariu@ni.com>
Co-authored-by: Natan Muntean <natan.muntean@ni.com>